Configure maven-release-plugin to play nice with Git
The release plugin normally pushes everything during a release. If
something goes wrong or you missed a setting, you have to release with a
new version if you don't want to rewrite/rebase history. Switching of
the auto-push behaviour allows to make informed choices on when to
actually push a release.